Scene 4, 5, 6, Cameras / Post Process Volume
Scene4_Camera added to *First* Matinee

Scene5and6_Camera added to *Second* Matinee (Note:  I messed up and
forgot to switch to my last camera for scene 6, so 5 and 6 are combined
into one camera.  This doesn't matter except for the organization)

Added a Post Process Volume that is floating above the flatform in the
shape of a cube.  I upped the saturation a ton, and turned up the
contrast a bit.  The colors are much more vivid and everything is less
washed out.  This item should be discussed as a group, but I think it
looks better than it did originally and at least semi-comic book-y.